requirements.txt
openai==2.14.0
requests==2.32.5 
python-dotenv==1.2.1
tabulate==0.9.0
httpie==3.2.4
fastapi==0.128.0
uvicorn==0.40.0
dotenv==0.9.9
google-genai==1.57.0